first try at SVE for ARM (with 512 bits register width as an hypothesis).
SVE (Scalable Vector Extension) is an upcoming SIMD/vector extension to the ARMv8 architecture.
Unlike previous SIMD extensions supported in FFTW3, the width of the vector registers is implementation-dependent.
This first draft hard-wire the value 512 bits for this width (4 double complex or 8 float complex in a register).

This is not yet meant to be used except for testing/evaluating SVE and associated tools.