Multi-width R5V, similar to arm-sve-clean.
This is a cleaned-up version of branch riscv-v from rdolbeau/fftw3.

No scalability, this is using multiple sets of masked operations to
support all width multiple of 128 bits.
This goes up to 16384 bits to support the EPAC VPU design.
Highly experimental, and still using the EPI-defined intrinsics, this
needs updating to the 'official' intirnsics for V1.0.
Contains some test code for alternate implemenations.